% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collect_res}
\alias{collect_res}
\title{collect values from a list of simulations}
\usage{
collect_res(res_list)
}
\arguments{
\item{res_list}{\emph{named} list of simulation results}
}
\description{
collect values from a list of simulations
}

