% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{do_extinct}
\alias{do_extinct}
\title{process extinction (remove zero-density strains)}
\usage{
do_extinct(state, strain)
}
\arguments{
\item{state}{state list}

\item{strain}{indices of strain(s) that went extinct}
}
\description{
process extinction (remove zero-density strains)
}

