% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{do_mut3}
\alias{do_mut3}
\title{do mutation, allowing for mutation in either trait}
\usage{
do_mut3(state, mut_var, mut_link, strain, mut_mean, mut_sd)
}
\arguments{
\item{state}{state list}

\item{mut_var}{which variable mutates}

\item{mut_link}{link/inv link function for traits}

\item{strain}{which strain(s) are mutating?}

\item{mut_mean}{mutation bias}

\item{mut_sd}{mutation SD}
}
\description{
do mutation, allowing for mutation in either trait
}

