% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{make_bilink}
\alias{make_bilink}
\title{make link functions for both parameters, based on tradeoff}
\usage{
make_bilink(tradeoff_fun, tradeoff_pars, gamma, beta, discrete = FALSE)
}
\arguments{
\item{tradeoff_fun}{string: "powerlaw" or "none"}

\item{tradeoff_pars}{listof parameters for tradeoff curve (g1, g2 = scale and curvature of power law)}

\item{gamma}{value of gamma}

\item{beta}{value of beta}

\item{discrete}{(logical)}
}
\description{
make link functions for both parameters, based on tradeoff
}

