% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{multlogit}
\alias{multlogit}
\title{Logit transform on (min,max) rather than (0,1);
reduces to log transform if max value is unbounded}
\usage{
multlogit(minval = 0, maxval = 1)
}
\arguments{
\item{minval}{minimum value}

\item{maxval}{maximum value}
}
\description{
Logit transform on (min,max) rather than (0,1);
reduces to log transform if max value is unbounded
}

