% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{run_sim}
\alias{run_sim}
\title{Evolutionary simulations}
\usage{
run_sim(inits = list(R0 = 2, gamma = 1/5, Ivec = NULL), mod_inits = list(),
  params = list(N = 1000, mu = 0.01, mut_type = "shift", mut_mean = -1, mut_sd
  = 0.5, mut_var = "beta", mut_link = NULL, tradeoff_pars = NULL),
  mod_params = list(), tradeoff_fun = "none", nt = 1e+05, dt = 1,
  rptfreq = max(nt/500, 1), discrete = TRUE, hazard = FALSE,
  seed = NULL, progress = FALSE, debug = FALSE, useCpp = FALSE)
}
\arguments{
\item{inits}{list
\itemize{
\item R0 (starting value of R0)
\item gamma (recovery rate)
\item Ivec (initial infectives)
}}

\item{mod_inits}{list of initial values to modify}

\item{params}{list: default values
\itemize{
\item N=1000 (population size)
\item mu (mutation probability per replication)
\item mut_type mutation model (only "shift" implemented)
\item mut_mean mean value of mutations
\item mut_sd mutational standard deviations
\item mut_var which variable(s) mutate?
\item mut_link link function for mutation
\item tradeoff_pars
}}

\item{mod_params}{list of parameters to modify}

\item{tradeoff_fun}{tradeoff function, "powerlaw" or "none"}

\item{nt}{number of time steps}

\item{dt}{within-step granularity}

\item{rptfreq}{reporting frequency (should divide nt)}

\item{discrete}{(logical) run discrete-time sim?}

\item{hazard}{parameterize events by hazard? (allows time-step cutting in discrete-time models)}

\item{seed}{random-number seed}

\item{progress}{draw progress bar?}

\item{debug}{(logical) debugging output?}

\item{useCpp}{(logical) use C++ code for continuous-time models?}
}
\description{
Evolutionary simulations
}

