% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{run_sim}
\alias{run_sim}
\title{Evolutionary simulations}
\usage{
run_sim(R0_init = 2, gamma0 = 1/5, gamma_max = Inf, N = 1000,
  mu = 0.01, mut_type = "shift", mut_mean = -1, mut_sd = 0.5,
  mut_var = "beta", mut_link = NULL, Ivec = NULL, nt = 1e+05,
  rptfreq = min(nt/500, 1), discrete = TRUE, seed = NULL,
  progress = FALSE, debug = FALSE, useCpp = FALSE)
}
\arguments{
\item{R0_init}{starting value of R0}

\item{gamma0}{recovery rate}

\item{gamma_max}{?? not used ??}

\item{N}{population size}

\item{mu}{mutation probability per replication}

\item{mut_type}{mutation model (only "shift" implemented)}

\item{mut_mean}{mean value of mutations}

\item{mut_sd}{mutational standard deviations}

\item{mut_var}{which variable(s) mutate?}

\item{mut_link}{link function for mutation}

\item{Ivec}{initial vector of infected numbers}

\item{nt}{number of time steps}

\item{rptfreq}{reporting frequency (should divide nt)}

\item{discrete}{(logical) run discrete-time sim?}

\item{seed}{random-number seed}

\item{progress}{draw progress bar?}

\item{debug}{(logical) debugging output?}

\item{useCpp}{(logical) use C++ code for continuous-time models?}
}
\description{
Evolutionary simulations
}

